+++ packages/client/src/clients/kamiden/client.ts
@@ -1,11 +1,10 @@
 import { grpc } from '@improbable-eng/grpc-web';
 import { createChannel, createClient } from 'nice-grpc-web';
 
-import { Feed, KamidenServiceClient, KamidenServiceDefinition, Message } from './types';
+import { KamidenServiceClient, KamidenServiceDefinition } from './proto';
+import { FeedCallbacks, MessageCallbacks } from './subscriptions';
 
 let Client: KamidenServiceClient | null = null;
-let messageCallbacks: ((message: Message) => void)[] = [];
-let feedCallbacks: ((feed: Feed) => void)[] = [];
 
 export function getClient(): KamidenServiceClient {
   if (!Client) {
@@ -16,44 +15,32 @@ export function getClient(): KamidenServiceClient {
     Client = createClient(KamidenServiceDefinition, channel);
 
     // Set up the perennial subscription
-    setupMessageSubscription();
+    setupSubscription();
   }
   return Client;
 }
 
-// Subscribe to messages and notify all registered callbacks
-async function setupMessageSubscription() {
+// Subscribe to messages and trigger all registered callbacks
+// NOTE(jb): not sure if callback handling should be the client's responsibility.
+// feels like it could get a bit messy
+async function setupSubscription() {
   try {
     const stream = Client!.subscribeToStream({});
 
     for await (const response of stream) {
       // Handle messages
       for (const message of response.Messages) {
-        messageCallbacks.forEach((callback) => callback(message));
+        MessageCallbacks.forEach((cb) => cb(message));
       }
 
       // Handle feed if present
       if (response.Feed) {
-        feedCallbacks.forEach((callback) => callback(response.Feed!));
+        FeedCallbacks.forEach((cb) => cb(response.Feed!));
       }
     }
   } catch (error) {
     console.error('[kamiden] Stream error:', error);
     // Attempt to reconnect after a delay
-    setTimeout(setupMessageSubscription, 5000);
+    setTimeout(setupSubscription, 5000);
   }
-}
-
-export function subscribeToMessages(callback: (message: Message) => void) {
-  messageCallbacks.push(callback);
-  return () => {
-    messageCallbacks = messageCallbacks.filter((cb) => cb !== callback);
-  };
-}
-
-export function subscribeToFeed(callback: (feed: Feed) => void) {
-  feedCallbacks.push(callback);
-  return () => {
-    feedCallbacks = feedCallbacks.filter((cb) => cb !== callback);
-  };
 }
