+++ packages/client/src/clients/kamiden/index.ts
@@ -1,1 +1,2 @@
 export { getClient as getKamidenClient } from './client';
+export { subscribeToFeed, subscribeToMessages } from './subscriptions';
